+++ soul-admin/src/main/java/org/dromara/soul/admin/model/query/RuleQuery.java
@@ -42,6 +42,11 @@ public class RuleQuery extends FilterQuery {
     private String selectorId;
 
     /**
+     * rule name.
+     */
+    private String name;
+
+    /**
      * page parameter.
      */
     private PageParameter pageParameter;