+++ soul-common/src/main/java/org/dromara/soul/common/enums/AdminPluginOperateEnum.java
@@ -45,6 +45,11 @@ public enum AdminPluginOperateEnum {
     EDIT("edit"),
 
     /**
+     * query operate.
+     */
+    QUERY("query"),
+
+    /**
      * modify operate.
      */
     SYNCHRONIZE("modify");